+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/Navbar/Navbar.tsx","./src/components/Task/AddEditTask.tsx","./src/components/Task/AddEditTaskWrapper.tsx","./src/components/Task/TaskFilter.tsx","./src/components/Task/TaskList.tsx","./src/components/Task/TaskListWrapper.tsx","./src/components/Task/TaskProvider.tsx","./src/components/Task/TaskStatusLegend.tsx","./src/types/Task/types.ts"],"version":"5.7.3"}
+{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/Navbar/Navbar.tsx","./src/components/NotFound/NotFound.tsx","./src/components/Task/AddEditTask.tsx","./src/components/Task/AddEditTaskWrapper.tsx","./src/components/Task/TaskFilter.tsx","./src/components/Task/TaskList.tsx","./src/components/Task/TaskListWrapper.tsx","./src/components/Task/TaskProvider.tsx","./src/components/Task/TaskStatusLegend.tsx","./src/types/Task/types.ts"],"version":"5.7.3"}
